pyproject.toml
[project]
name = "reinforcement-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gymnasium>=1.2.2",
    "metaworld>=3.0.0",
    "moviepy>=2.2.1",
    "numpy>=2.3.5",
    "packaging>=25.0",
    "pyaml>=25.7.0",
    "torch>=2.9.1",
]
